v1.3.0 -> Alterações e Adições, Double para Int -->Grafo { >Alteração no tipo de dados contido na matriz de adjacência >de inteiro para double } -->Array<Array<String>>.replaceInfinity() { >Função de extensão para matrizes de string que converte todos >os valores iguais à constante INFINITO para o símbolo unicode >de infinito } -->ViewUtil { >Singleton removido e fontes agora são membros top level >Função replaceInfinity convertido para função de extensão } -->aplicacao { >Todos os singletons contendo a função estática main() >foram removidos e a função foi convertida para top level }

+++ src/main/kotlin/teoriadosgrafos/Aresta.kt
@@ -7,7 +7,7 @@ import java.util.Objects
  *
  * @author [Icaro D Temponi](https://github.com/icarohs7)
  */
-data class Aresta(val origem: Int, val destino: Int, val peso: Int) {
+data class Aresta(val origem: Int, val destino: Int, val peso: Double) {
 	
 	/**
 	 * Hash code int.
@@ -32,8 +32,8 @@ data class Aresta(val origem: Int, val destino: Int, val peso: Int) {
 			return false
 		}
 		return (origem == other.origem && destino == other.destino
-		        || destino == other.origem && origem == other.destino)
-		       && peso == other.peso
+				|| destino == other.origem && origem == other.destino)
+				&& peso == other.peso
 	}
 	
 	/**
